Update Model
Changed card model to use foreign key to the Card_Set model. Allows for easier linking of the sets and card and makes handling searching for cards based on set easier to handle.

Also imported all current cards from the API to the database so info can be taken from the local database for hopefully faster results.